# dbr quant overhead[10/x]: disable torch_function overrides for leaf nodes