--- a/torch/ao/quantization/_dbr/auto_trace.py
+++ b/torch/ao/quantization/_dbr/auto_trace.py
@@ -58,6 +58,8 @@ def add_auto_observation(
     # logging.
     global_op_idx = [0]
 
+    global_disable_torch_function_override = False
+
     class QuantizationPrepareTensorProxy(torch.Tensor):
         """
         An override of `torch.Tensor` to enable dynamic tracing for
@@ -79,6 +81,10 @@ def add_auto_observation(
 
         @classmethod
         def __torch_function__(cls, func, types, args=(), kwargs=None):
+            nonlocal global_disable_torch_function_override
+            if global_disable_torch_function_override:
+                return super().__torch_function__(func, types, args, kwargs)
+
             # to prevent printing things from going into an infinite loop
             if func == torch.Tensor.__repr__:
                 return super().__torch_function__(func, types, args, kwargs)
@@ -188,12 +194,25 @@ def add_auto_observation(
                         # before hooks
                         if not first_call:
                             parent_qstate.validate_cur_op(cur_module)
+
+                        # If we are in this hook, `cur_module` is a leaf module.
+                        # Therefore, we do not need to override any of its
+                        # children. Disabling the overrides for performance.
+                        nonlocal global_disable_torch_function_override
+                        old_global_disable_torch_function_override = \
+                            global_disable_torch_function_override
+                        global_disable_torch_function_override = True
+
                         args, kwargs = parent_qstate.op_prepare_before_hook(
                             cur_module, args, kwargs, first_call, qtensor_id,
                             fqn, cur_module)
 
                         # original forward
                         output = orig_module_call(self, *args, **kwargs)
+
+                        # Re-enable the overrides.
+                        global_disable_torch_function_override = \
+                            old_global_disable_torch_function_override
 
                         # after hooks
                         # TODO is it correct to call_cur_module twice here?
@@ -294,6 +313,8 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
     # logging.
     global_op_idx = [0]
 
+    global_disable_torch_function_override = False
+
     class QuantizationConvertTensorProxy(torch.Tensor):
         """
         An override of `torch.Tensor` to enable dynamic dispatch for
@@ -316,6 +337,10 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
 
         @classmethod
         def __torch_function__(cls, func, types, args=(), kwargs=None):
+            nonlocal global_disable_torch_function_override
+            if global_disable_torch_function_override:
+                return super().__torch_function__(func, types, args, kwargs)
+
             # to prevent printing things from going into an infinite loop
             if func == torch.Tensor.__repr__:
                 return super().__torch_function__(func, types, args, kwargs)
@@ -438,11 +463,12 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
                     module_stack.append(self)
                     hook_type = get_module_hook_type(parent_module, cur_module)
                     if enable_logging:
-                        logger.debug(
-                            f"_patched_module_call {type(self)} " +
-                            # f"arg_types {[type(arg) for arg in args]} " +
-                            f"arg_dtypes {[arg.dtype if isinstance(arg, torch.Tensor) else None for arg in args]} " +
-                            f"hook_type {hook_type}")
+                        with torch._C.DisableTorchFunction():
+                            logger.debug(
+                                f"_patched_module_call {type(self)} " +
+                                # f"arg_types {[type(arg) for arg in args]} " +
+                                f"arg_dtypes {[arg.dtype if isinstance(arg, torch.Tensor) else None for arg in args]} " +
+                                f"hook_type {hook_type}")
 
                     if hook_type is HookType.OP_HOOKS:
                         # before hooks
@@ -451,6 +477,15 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
                         if enable_logging:
                             logger.debug(qstate)
                         qstate.validate_cur_op(cur_module)
+
+                        # If we are in this hook, `cur_module` is a leaf module.
+                        # Therefore, we do not need to override any of its
+                        # children. Disabling the overrides for performance.
+                        nonlocal global_disable_torch_function_override
+                        old_global_disable_torch_function_override = \
+                            global_disable_torch_function_override
+                        global_disable_torch_function_override = True
+
                         _, args, kwargs = qstate.op_convert_before_hook(
                             cur_module, args, kwargs, cur_module)
                         # forward
@@ -458,6 +493,11 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
                         # after hooks
                         output = qstate.op_convert_after_hook(
                             cur_module, output, global_op_idx)
+
+                        # Re-enable the override.
+                        global_disable_torch_function_override = \
+                            old_global_disable_torch_function_override
+
                         qstate.mark_cur_op_complete(cur_module)
 
                     elif hook_type is HookType.MODULE_IO_HOOKS:
@@ -494,12 +534,13 @@ def add_auto_convert(module : torch.nn.Module) -> torch.nn.Module:
                         output = orig_module_call(self, *args, **kwargs)
 
                     if enable_logging:
-                        logger.debug(
-                            f"_patched_module_call {type(self)} " +
-                            # f"out {type(output)} " +
-                            f"dtype {output.dtype if isinstance(output, torch.Tensor) else None} " +
-                            "end")
-                        logger.debug(f"ending fqn {fqn}\n")
+                        with torch._C.DisableTorchFunction():
+                            logger.debug(
+                                f"_patched_module_call {type(self)} " +
+                                # f"out {type(output)} " +
+                                f"dtype {output.dtype if isinstance(output, torch.Tensor) else None} " +
+                                "end")
+                            logger.debug(f"ending fqn {fqn}\n")
                     return output
                 finally:
                     module_stack.pop()