Stack from [ghstack](https://github.com/ezyang/ghstack):
* #69002 [not for land] dbr quant: turn off overrides on regions of model
* #69001 [not for land] dbr quant: add any_child_needs_op_hooks flag
* #69000 [not for land] dbr quant: add any_child_needs_arg_dequants flag
* #68999 [not for lnad] dbr quant: improve needs_dtype_transform_on_outputs
* #68877 dbr quant overhead [14/x]: cache whether an op is a module
* #68841 dbr quant overhead [13/x]: cache results of get_module_hook_type
* #68840 dbr quant: fix debugging fqn info for converted model
* #68839 dbr quant overhead[12/x]: turn off overrides for module convert output hook
* #68838 dbr quant: remove unnecessary outputs hook
* #68837 dbr quant overhead[11/x]: speed up module convert hook
* **#68836 dbr quant overhead[10/x]: disable torch_function overrides for leaf nodes**

Summary:

If we have a leaf module like a `torch.nn.Conv2d`, DBR quant handles
the input and output of the module and should treat the inside of
this module as invisible.  Specifically, there is no need to override
the `F.conv2d` call if the parent module is already being overridden.

Before this PR, `__torch_function__` was still overridden for the insides
of leaf modules, and the override was a no-op.  There was some overhead
in these overrides because they were checking the hook type.

This PR adds a fast global override so we can skip overridding the insides
of leaf modules. This has some performance benefits in the prepare model,
because we now skip overriding all of the inner functions in observers.

Test plan:

testing
```
python test/test_quantization.py TestQuantizeDBR
```

perf
```
// MobileNetV2, 1x3x224x224, comparing fp32 with dbr quant, Mac OS laptop

// before

fp32: 0.017837 seconds avg
fx_prepared: 0.021963 seconds avg, 0.812143 speedup vs fp32
fx_quantized: 0.012632 seconds avg, 1.412056 speedup vs fp32
dt_prepared: 0.034052 seconds avg, 0.523820 speedup vs fp32
dt_quantized: 0.018316 seconds avg, 0.973829 speedup vs fp32

// after

fp32: 0.020395 seconds avg
fx_prepared: 0.026969 seconds avg, 0.756230 speedup vs fp32
fx_quantized: 0.013195 seconds avg, 1.545611 speedup vs fp32
dt_prepared: 0.033432 seconds avg, 0.610023 speedup vs fp32
dt_quantized: 0.018244 seconds avg, 1.117866 speedup vs fp32

```

Differential Revision: [D32630883](https://our.internmc.facebook.com/intern/diff/D32630883)